Add QueryParams normalization and generation (#2)
FOLIO endpoints are fairly consistent in their handling of query and sort, relying on CQL. ERM endpoints follow different rules (there are probably others and support will be added as they are found). LDLite (and really any folio client) relies on having one parameter for http requests. This is mostly fine if we weren't also trying to do some magic manipulations of limits, offsets, and ids to page or get single records. I've pulled out the normalization logic from LDLite and expanded upon it with many many tests and a couple of bug fixes here. It's complicated but FOLIO is complicated.